refactor(resource-interact): 修复阅读量/点赞/打分功能 PR 评审问题
- 移除 document/note 服务中不必要的 recordRead 远程 Feign 调用
- 删除 ResourceReadRecordReqDTO、ResourceInteractStateResponse 及相关接口定义
- 消除并发双击问题，点赞/评分改用 upsert 原子写保证幂等
- 互动信息计数字段统一由 Long 改为 Integer，消除不必要精度放大
- toggleLike/rateResource 返回类型改为 void
- 消除 orElse(null) 空对象兜底，新增 ResourceInteractInfoEntity 无参与含参构造器
- 提取 Redis key 前缀常量至 RedisCacheManager，消除魔法字符串

+++ wisepen-resource-service/wisepen-resource-api/src/main/java/com/oriole/wisepen/resource/domain/dto/res/ResourceItemResponse.java
@@ -13,8 +13,8 @@ import java.util.Map;
 @Data
 public class ResourceItemResponse extends ResourceItemInfoBase {
     private String resourceId;
-    private Long readCount;   // 聚合自资源互动信息表，缺失时返回 0
-    private Long likeCount;   // 聚合自资源互动信息表，缺失时返回 0
+    private Integer readCount; // 聚合自资源互动信息表，缺失时返回 0
+    private Integer likeCount; // 聚合自资源互动信息表，缺失时返回 0
     private Double scoreAvg;  // 聚合自资源互动信息表，无评分时为 null
     private Boolean liked;      // 当前用户点赞状态；无用户上下文时可为 null
     private Integer userScore;  // 当前用户评分；未评分时为 null
